Relocated helpers into appropriate files
