+++ src/api/bot/user/roles/delete.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express-serve-static-core";
-import { GetGuildUser } from "../../../../common/discord";
+import { GetGuildUser, GetDiscordUser } from "../../../../common/helpers/discord";
 import { Role } from "discord.js";
-import { genericServerError, GetDiscordUser, getProjectsByUserDiscordId } from "../../../../common/helpers";
+import { genericServerError } from "../../../../common/helpers/generic";
 
 module.exports = async (req: Request, res: Response) => {
     if (!req.headers.authorization) {
